Improved naming of test files.
