+++ modua/modua_app/test_views.py
@@ -16,101 +16,7 @@ class TestViews(APITestCase):
 
 
     def setUp(self):
-        Definitions.objects.create(word_character="hello",
-                                           definition="A greeting")
-        Definitions.objects.create(word_character='kool-aid',
-                                   definition='the koolest drink ever')
-        User.objects.create_user('john', 'john@gmail.com', 'password')
-
-
-    def test_search_status_code_200(self):
-        '''
-        Passes if a valid search returns a status code of 200.
-        '''
-        response = self.client.get("/api/0.1/search/en-US/hello", format='json')
-        self.assertEqual(response.status_code, 200)
-
-    def test_term_search(self):
-        '''
-        Passes if the json returned from a valid request is the correct json.
-        '''
-        response = self.client.get("/api/0.1/search/en-US/hello", format='json')
-        json = response.json()
-        self.assertEqual(json, {'word_character': "hello",
-                                    'definition': 'A greeting',
-                                    'transliteration': None })
-    
-    def test_bad_term_search(self):
-        '''
-        Passes if a request for a word not in the dictionary returns a status code 404.
-        '''
-        response = self.client.get("/api/0.1/search/en-US/term_not_in_DB", format='json')
-        self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
-    
-    def test_bad_url(self):
-        '''
-        Passes if an invalid url returns a 404 response.
-        '''
-        response = self.client.get("/api/badurl", format='json')
-        self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
-
-    def test_nondelimited_term_search(self):
-        '''
-        Passes if a the json returned from a valid search for a nondelimited word is correct.
-        '''
-        response = self.client.get("/api/0.1/search/zh-Hant/kool-aidxxxx", format='json')
-        json = response.json()
-        self.assertEqual(json, {'word_character': 'kool-aid',
-                                    'definition': 'the koolest drink ever',
-                                    'transliteration': None
-                               })
-    
-    def test_bad_nondelimited_search(self):
-        '''
-        Passes if a request for a nondelimited word that isn't in the database returns a 404 response.
-        '''
-        response = self.client.get("/api/0.1/search/zh-Hant/term_not_in_DB", format='json')
-        self.assertEqual(response.status_code, status.HTTP_404_NOT_FOUND)
-
-    def test_user_search_status_code_200(self):
-        '''
-        Passes if a valid search by a registered user returns a status code of 200.
-        '''
-        self.client.login(username='john', password='password')
-        response = self.client.get("/api/0.1/search/en-US/hello", format='json')
-        self.assertEqual(response.status_code, 200)
-
-    def test_user_valid_search(self):
-        '''
-        Passes if registered user gets a correct json response for a valid search.
-        '''
-        self.client.login(username='john', password='password')
-        response = self.client.get("/api/0.1/search/en-US/hello", format='json')
-        json = response.json()
-        self.assertEqual(json, {'word_character': "hello",
-                                    'definition': 'A greeting',
-                                    'transliteration': None })
-
-    def test_language_list_one_language(self):
-        '''
-        Passes if /languages/ returns the correct json.
-        '''
-        Languages.objects.create(language='en')
-        response = self.client.get("/api/0.1/languages/", format='json')
-        json = response.json()[0]
-        self.assertEqual(json, {'language': 'en'})
-
-
-    def test_language_list_two_languages(self):
-        '''
-        Passes if /languages/ returns the correct json.
-        '''
-        Languages.objects.create(language='en')
-        Languages.objects.create(language='zh')
-        response = self.client.get("/api/0.1/languages/", format='json')
-        json = response.json()
-        self.assertEqual(json, [{'language': 'en'},
-                                {'language': 'zh'}])
+        pass
 
 
     def test_language_wordlist_view_only(self):
@@ -142,7 +48,7 @@ class TestViews(APITestCase):
                       'definition': 'godly saying',
                       'transliteration': None 
                   }
-                 ],
+                ],
             key=lambda t:t['word_character']
         )
